@@ -6,7 +6,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"os"
 	"path"
 	"strings"
@@ -99,7 +98,7 @@ func (b TestRunner) exportJunitReport(opts types.RunTestOptions, stdout string, 
 
 	junitFilename := path.Join(testRunner.ReportsDirectory,
 		fmt.Sprintf("junit-%s.xml", strings.ReplaceAll(opts.ImageName, "/", "_")))
-	if err := ioutil.WriteFile(junitFilename, []byte(stdout), 0o644); err != nil { // nolint: gosec
+	if err := os.WriteFile(junitFilename, []byte(stdout), 0o644); err != nil { //nolint:gosec
 		return fmt.Errorf("could not write junit report to file %s: %w", junitFilename, err)
 	}
 	return nil
