@@ -91,8 +91,8 @@ func renderConsoleOutput(imagesList []dag.Image) {
 	table := tablewriter.NewWriter(os.Stdout)
 	table.SetAutoWrapText(false)
 	table.SetAutoFormatHeaders(true)
-	table.SetHeaderAlignment(tablewriter.ALIGN_LEFT)
-	table.SetAlignment(tablewriter.ALIGN_LEFT)
+	table.SetHeaderAlignment(tablewriter.ALIGN_LEFT) //nolint:nosnakecase
+	table.SetAlignment(tablewriter.ALIGN_LEFT)       //nolint:nosnakecase
 	table.SetCenterSeparator("")
 	table.SetColumnSeparator("")
 	table.SetRowSeparator("")
