@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-// nolint:paralleltest
+//nolint:paralleltest
 func Test_DGossExecutor_NewDGossExecutorUsesDefaultShell(t *testing.T) {
 	err := os.Unsetenv("SHELL")
 	require.NoError(t, err)
@@ -19,7 +19,7 @@ func Test_DGossExecutor_NewDGossExecutorUsesDefaultShell(t *testing.T) {
 	assert.Equal(t, "/bin/bash", executor.Shell)
 }
 
-// nolint:paralleltest
+//nolint:paralleltest
 func Test_DGossExecutor_NewDGossExecutorDetectsShellFromEnv(t *testing.T) {
 	t.Setenv("SHELL", "/path/to/shell")
 
